refactor(phase-4): split spotifyext into extensions/spotify/ package
Extract the 1375-line ``extensions/spotifyext.py`` god-class into a mixin-based
package and move the pure Spotify API layer out of ``src/`` into a new
``src/integrations/`` tree:

- ``src/integrations/spotify.py`` (81 lines): pure Spotify helpers
  (``spotify_auth``, ``spotifymongoformat``) with no Discord imports.
- ``extensions/spotify/_common.py`` (345 lines): shared state — ``ServerData``,
  ``VoteManager``, ``EmbedType``, ``SERVERS`` dict, shared ``sp`` client, and
  the Discord embed builders that previously lived in ``src/spotify.py``.
- ``extensions/spotify/auth.py`` (67 lines): ``AuthMixin`` — ``/updatetoken``.
- ``extensions/spotify/playlist.py`` (442 lines): ``PlaylistMixin`` —
  ``/addsong``, ``/songinfo``, autocompletes, ``check_playlist_changes`` and
  ``new_titles_playlist`` tasks.
- ``extensions/spotify/votes.py`` (680 lines): ``VotesMixin`` —
  ``/rappelvote``, ``/addwithvote``, ``/nextvote``, the two button handlers,
  ``randomvote`` / ``reminder_check`` / ``check_for_end`` tasks, ``endvote``.
- ``extensions/spotify/__init__.py`` (85 lines): the glue ``SpotifyExtension``
  class composing the mixins and hosting lifecycle / MongoDB state helpers.

Deletes the monolithic ``extensions/spotifyext.py`` and ``src/spotify.py``. The
new layout keeps decorated methods on mixins visible to ``interactions.py``
extension discovery via MRO, matching the pattern used for the twitch split
(PR #21).
