+++ src/integrations/__init__.py
@@ -1,0 +1,1 @@
+"""Pure external-API integration modules (no Discord imports)."""
